Move `TemporaryLocalDirectory` to the test support library.
