+++ tiledb/sm/filesystem/CMakeLists.txt
@@ -35,7 +35,6 @@ commence(object_library vfs)
         mem_filesystem.cc
         path_win.cc
         posix.cc
-        temporary_local_directory.cc
         uri.cc
         vfs.cc
         vfs_file_handle.cc
